+++ apps/explorer/lib/explorer/smart_contract/verifier/constructor_arguments.ex
@@ -1,3 +1,4 @@
+# credo:disable-for-this-file
 defmodule Explorer.SmartContract.Verifier.ConstructorArguments do
   @moduledoc """
   Smart contract contrstructor arguments verification logic.
@@ -10,7 +11,7 @@ defmodule Explorer.SmartContract.Verifier.ConstructorArguments do
   @metadata_hash_prefix_0_5_11 "a265627a7a72315820"
   @metadata_hash_prefix_0_6_0 "a264697066735822"
 
-  @metadata_hash_common_suffix "64736f6c6343"
+  @metadata_hash_common_suffix "64736f6c63"
 
   def verify(address_hash, contract_code, arguments_data, contract_source_code, contract_name) do
     arguments_data = arguments_data |> String.trim_trailing() |> String.trim_leading("0x")
@@ -53,7 +54,51 @@ defmodule Explorer.SmartContract.Verifier.ConstructorArguments do
       # Solidity >= 0.5.10 https://solidity.readthedocs.io/en/v0.5.10/metadata.html
       @metadata_hash_prefix_0_5_10 <>
           <<_::binary-size(64)>> <>
-          @metadata_hash_common_suffix <> <<_::binary-size(6)>> <> "0032" <> constructor_arguments ->
+          @metadata_hash_common_suffix <> "43" <> <<_::binary-size(6)>> <> "0032" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_10
+        )
+
+      @metadata_hash_prefix_0_5_10 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7826" <> <<_::binary-size(76)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_10
+        )
+
+      @metadata_hash_prefix_0_5_10 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7827" <> <<_::binary-size(78)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_10
+        )
+
+      @metadata_hash_prefix_0_5_10 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7828" <> <<_::binary-size(80)>> <> "0058" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_10
+        )
+
+      @metadata_hash_prefix_0_5_10 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7829" <> <<_::binary-size(82)>> <> "0059" <> constructor_arguments ->
         split_constructor_arguments_and_extract_check_func(
           constructor_arguments,
           check_func,
@@ -66,7 +111,51 @@ defmodule Explorer.SmartContract.Verifier.ConstructorArguments do
       # Metadata: Update the swarm hash to the current specification, changes bzzr0 to bzzr1 and urls to use bzz-raw://
       @metadata_hash_prefix_0_5_11 <>
           <<_::binary-size(64)>> <>
-          @metadata_hash_common_suffix <> <<_::binary-size(6)>> <> "0032" <> constructor_arguments ->
+          @metadata_hash_common_suffix <> "43" <> <<_::binary-size(6)>> <> "0032" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_11
+        )
+
+      @metadata_hash_prefix_0_5_11 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7826" <> <<_::binary-size(76)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_11
+        )
+
+      @metadata_hash_prefix_0_5_11 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7827" <> <<_::binary-size(78)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_11
+        )
+
+      @metadata_hash_prefix_0_5_11 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7828" <> <<_::binary-size(80)>> <> "0058" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_5_11
+        )
+
+      @metadata_hash_prefix_0_5_11 <>
+          <<_::binary-size(64)>> <>
+          @metadata_hash_common_suffix <> "7829" <> <<_::binary-size(82)>> <> "0059" <> constructor_arguments ->
         split_constructor_arguments_and_extract_check_func(
           constructor_arguments,
           check_func,
@@ -87,7 +176,51 @@ defmodule Explorer.SmartContract.Verifier.ConstructorArguments do
       # Fixing PR has been created https://github.com/ethereum/solidity/pull/8174
       @metadata_hash_prefix_0_6_0 <>
           <<_::binary-size(68)>> <>
-          @metadata_hash_common_suffix <> <<_::binary-size(6)>> <> "0033" <> constructor_arguments ->
+          @metadata_hash_common_suffix <> "43" <> <<_::binary-size(6)>> <> "0033" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_6_0
+        )
+
+      @metadata_hash_prefix_0_6_0 <>
+          <<_::binary-size(68)>> <>
+          @metadata_hash_common_suffix <> "7826" <> <<_::binary-size(76)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_6_0
+        )
+
+      @metadata_hash_prefix_0_6_0 <>
+          <<_::binary-size(68)>> <>
+          @metadata_hash_common_suffix <> "7827" <> <<_::binary-size(78)>> <> "0057" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_6_0
+        )
+
+      @metadata_hash_prefix_0_6_0 <>
+          <<_::binary-size(68)>> <>
+          @metadata_hash_common_suffix <> "7828" <> <<_::binary-size(80)>> <> "0058" <> constructor_arguments ->
+        split_constructor_arguments_and_extract_check_func(
+          constructor_arguments,
+          check_func,
+          contract_source_code,
+          contract_name,
+          @metadata_hash_prefix_0_6_0
+        )
+
+      @metadata_hash_prefix_0_6_0 <>
+          <<_::binary-size(68)>> <>
+          @metadata_hash_common_suffix <> "7829" <> <<_::binary-size(82)>> <> "0059" <> constructor_arguments ->
         split_constructor_arguments_and_extract_check_func(
           constructor_arguments,
           check_func,