+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ### Features
 
 ### Fixes
-- [#3106](https://github.com/poanetwork/blockscout/pull/3106) - Verify contract using creation tx input
+- [#3112](https://github.com/poanetwork/blockscout/pull/3112) - Fix verification of contracts, compiled with nightly builds of solc compiler
+- [#3112](https://github.com/poanetwork/blockscout/pull/3112) - Check compiler version at contract verification
+- [#3106](https://github.com/poanetwork/blockscout/pull/3106) - Fix verification of contracts with `immutable` declaration
+- [#3106](https://github.com/poanetwork/blockscout/pull/3106) - Fix verification of contracts, created from factory (from internal transaction)
 
 ### Chore